Add tests. Fix everything

+++ routers/posts.py
@@ -14,21 +14,21 @@ from configs.config_activity import ActionType
 router = APIRouter()
 
 class PostTextContent(BaseModel):
-    title: Annotated[str, Query(min_length=1, max_length=100)]
-    content: Annotated[str, Query(min_length=1)]
-    tag: Annotated[str, Query(min_length=1)]
+    title: str
+    content: str
+    tag: str
 
     @classmethod
     def form(
         cls,
-        title: Annotated[str, Form()],
-        content: Annotated[str, Form()],
-        tag: Annotated[str, Form()]
+        title: Annotated[str, Form(min_length=1)],
+        content: Annotated[str, Form(min_length=1)],
+        tag: Annotated[str, Form(min_length=1)]
     ):
         return cls(title=title, content=content, tag=tag)
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def get_newsfeed(this_user: User_auth, criteria: str, offset: int = 0, limit: int =15):
+async def get_newsfeed(this_user: User_auth, criteria: str | None = None, offset: int = 0, limit: int = 15):
     """
     Get latest posts for user's feed.\n
     Return a list of post_id. To retrieve their content, make GET request for each post.
@@ -42,7 +42,7 @@ async def get_post(post_id: int, this_user: User_auth, db: Db_dependency):
     """
     Get the post by post_id
     """
-    post = await postutils.getOutputPost(this_user.user_id, post_id, db)
+    post = await postutils.getOutputPost(this_user, post_id, db)
     if post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     return post
@@ -97,7 +97,10 @@ async def edit_post(
     # Update post content
     post = await postutils.getPost(post_id, db)
     if post is None:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="This action is not allowed")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    
+    if post.author_id != this_user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not permitted")
     
     await postutils.updatePost(db, post, text_content.title, text_content.content, text_content.tag)
 
@@ -132,7 +135,10 @@ async def delete_post(this_user: User_auth, post_id: int, db: Db_dependency):
     """
     post = await postutils.getPost(post_id, db)
     if post is None:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="This action is not allowed")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+
+    if post.author_id != this_user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not permitted")
     
     await postutils.deletePost(db, post)
 
@@ -141,17 +147,11 @@ async def delete_post(this_user: User_auth, post_id: int, db: Db_dependency):
     }
 
 @router.post("/posts/{post_id}/vote", status_code=status.HTTP_200_OK)
-async def vote_post(this_user: User_auth, post_id: int, vote_type: int, db: Db_dependency):
+async def vote_post(this_user: User_auth, post_id: int, vote_type: Annotated[int, Form()], db: Db_dependency):
     """
     Change user's vote of a post
     Vote type can be -1, 0, 1 for downvote, no vote or upvote
     """
-
-    VOTE_TYPE = ["novote", "upvote", "downvote"]
-
-    # Check if value is valid
-    if vote_type not in [-1, 0, 1]:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid value")
 
     # Check if post exists
     post = await postutils.getPost(post_id, db)
@@ -159,18 +159,7 @@ async def vote_post(this_user: User_auth, post_id: int, vote_type: int, db: Db_d
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
     # Update vote count
-    vote = post.votes.filter(PostVote.user_id == this_user.user_id).first()
-    if vote is None:
-        vote = PostVote(user_id=this_user.user_id, post_id=post_id, value=0)
-        db.add(vote)
-        db.commit()
-        db.refresh(vote)
-
-        # If this action is new, log the action
-        await logActivity(this_user.user_id, db, ActionType.VOTECOMMENT, VOTE_TYPE[vote_type], vote.vote_id, post.author_id)
-        
-    post.vote_count += vote_type - vote.value
-    vote.value = vote_type
-    db.commit()
+    if not await postutils.votePost(db, this_user, post, vote_type):
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid value")
 
     return {"message": "Voted"}
